feat: add lambda function that can induce load on the service (#2)
This PR adds a lambda function to the infrastructure that can induce
artificial load on the example service so that the deployment can be
tested with load on the system.

Note this could create some extra cost. The Lambda is triggered by a
cloudwatch events schedule which can be disabled if wanted. That way,
users have to opt-in to this new behavior